--- a/v1/topdown/strings_bench_test.go
+++ b/v1/topdown/strings_bench_test.go
@@ -113,3 +113,36 @@ func generateBulkStartsWithInput() map[string]interface{} {
 		"prefixes": prefixes,
 	}
 }
+
+func BenchmarkSplit(b *testing.B) {
+	bctx := BuiltinContext{}
+	operands := []*ast.Term{
+		ast.StringTerm("a.b.c.d.e"),
+		ast.StringTerm("."),
+	}
+
+	exp := eqIter(ast.ArrayTerm(
+		ast.StringTerm("a"),
+		ast.StringTerm("b"),
+		ast.StringTerm("c"),
+		ast.StringTerm("d"),
+		ast.StringTerm("e"),
+	))
+
+	b.ResetTimer()
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		if err := builtinSplit(bctx, operands, exp); err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func eqIter(a *ast.Term) func(*ast.Term) error {
+	return func(b *ast.Term) error {
+		if !a.Equal(b) {
+			return fmt.Errorf("expected %v equal to %v", a, b)
+		}
+		return nil
+	}
+}