--- a/v1/util/performance_test.go
+++ b/v1/util/performance_test.go
@@ -1,0 +1,17 @@
+package util
+
+import "testing"
+
+type testStruct struct {
+	foo int
+}
+
+func BenchmarkNewPtrSlice(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		s := NewPtrSlice[testStruct](100)
+		for j := 0; j < 100; j++ {
+			s[j].foo = j
+		}
+	}
+}